on a second thought, we should not mix floating point with integer arithmetic.

+++ JNF_NEAT/neural_network_trainer.cpp
@@ -37,7 +37,7 @@ void NeuralNetworkTrainer::SetBodies(std::vector<IBody*>& bodies) {
 }
 
 
-void NeuralNetworkTrainer::TrainUntilFitnessEquals(double fitnessToReach) {
+void NeuralNetworkTrainer::TrainUntilFitnessEquals(size_t fitnessToReach) {
 	LetGenerationLive();
 	while (GetFittestOrganism().GetOrCalculateRawFitness() < fitnessToReach) {
 		Repopulate();
